requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-chroma
chromadb
python-dotenv
pypdf2 
langchain-core 
langchain-community 
langchain-text-splitters
langchain-classic
langchain-ollama